Did stylings for for FitnessForm and NutritionForm components. Created NutritionList component and filled out more of FitnessList component and did stylings for both. Did styling for profile page.
