+++ client/src/pages/Profile.jsx
@@ -1,33 +1,19 @@
-import { Navigate, useParams } from 'react-router-dom';
 import { useQuery } from '@apollo/client';
 
-// import ThoughtForm from '../components/ThoughtForm';
-// import ThoughtList from '../components/ThoughtList';
 import FitnessForm from '../components/FitnessForm';
-import FoodForm from '../components/FoodForm';
+import NutritionForm from '../components/NutritionForm';
 import FitnessList from '../components/FitnessList';
+import NutritionList from '../components/NutritionList';
 
 import { QUERY_USER, QUERY_ME } from '../utils/queries';
 
 import Auth from '../utils/auth';
 
 const Profile = () => {
-  // const { username: userParam } = useParams();
-
-  // const { loading, data } = useQuery(userParam ? QUERY_USER : QUERY_ME, {
-  //   variables: { username: userParam },
-  // });
 
   const { loading, data } = useQuery(QUERY_ME)
 
   const user = data?.me || {};
-  // if (
-  //   Auth.loggedIn() && 
-  //   /* Run the getProfile() method to get access to the unencrypted token value in order to retrieve the user's username, and compare it to the userParam variable */
-  //   Auth.getProfile().authenticatedPerson.username === userParam
-  // ) {
-  //   return <Navigate to="/me" />;
-  // }
 
   if (loading) {
     return <div>Loading...</div>;
@@ -43,25 +29,31 @@ const Profile = () => {
   }
 
   return (
-    <div>
+    <div className='flex-column'>
       <h2>
-        This is {user.usernames}'s Profile Page
+        This is {user.username}'s Profile Page
       </h2>
       <div>
         {/* <LineGraph />
         <LineGraph /> */}
       </div>
-      <div>
+      <div className='flex-row my-3'>
         <FitnessForm workouts={user.workouts}/>
-        <FoodForm foodIntake={user.foodIntake}/>
+        <NutritionForm foodIntake={user.foodIntake}/>
       </div>
       <div>
-        <div>
-          <div>
+        <div className='flex-row'>
+          <div className='col-12 col-lg-6'>
             <h3>
-              Workout Data for {user.userName}
+              Workout Data for {user.username}
             </h3>
             <FitnessList workouts={user.workouts} />
+          </div>
+          <div className='col-12 col-lg-6'>
+            <h3>
+              Nutrition Data for {user.username}
+            </h3>
+            <NutritionList foodIntake={user.foodIntake} />
           </div>
         </div>
       </div>
